\name{AccountCombiner}
\alias{AccountCombiner}
\title{
Combine Accounts within a Dataset
}
\description{
Combine the stats from accounts if a player is playing on an alternate account in the given replay group data (generated by stats_from_group).
}
\usage{
AccountCombiner(data, main, alt1, alt2 = NULL, alt3 = NULL)
}
\arguments{
  \item{data}{
Dataset from the stats_from_group() function.
}
  \item{main}{
The player's main account ID.
}
  \item{alt1}{
Player's first alternate account ID.
}
  \item{alt2}{
Player's second alternate account ID (optional).
}
  \item{alt3}{
Player's third alternate account ID (optional).
}
}

\value{
Returns a dataframe of player stats from the replay group with combined accounts, in the same format as the input.
}
\references{
Ballchasing API Documentation- https://ballchasing.com/doc/api \cr
Thanks to CantFly for creating Ballchasing.
}
\author{
Brett Jacot
}
\note{
Note that I've changed a few digits from my own API key in the example code, as API keys are private and user-specific.
}


\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or standard data sets, see data().

## The function is currently defined as
function (data, main, alt1, alt2 = NULL, alt3 = NULL)
{
    alts <- Filter(Negate(is.null), list(alt1, alt2, alt3))
    if (length(alts) == 0)
        return(data)
    fixDF <- data[data$id \%in\% c(main, alts), ]
    print(str(fixDF))
    data2 <- data[!data$id \%in\% c(main, alts), ]
    print(str(data2))
    combined <- fixDF \%>\% summarise(id = main, across(starts_with(c("game_average",
        "cumulative.win_percentage")), ~weighted.mean(.x, fixDF$cumulative.games,
        na.rm = TRUE)), across(c("cumulative.games", "cumulative.wins"),
        sum, na.rm = TRUE))
    print(str(combined))
    combined$game_average.core.shooting_percentage <- (combined$game_average.core.goals/combined$game_average.core.shots) *
        100
    final <- rbind(data2, combined)
    print(str(final))
    return(final)
  }



OCCBStats_noname <- stats_from_group(group_id = "ontario-campus-carball-winter-20-xxukviq8ww", api_key = P9LtEgF40vIUHUcrPrD4p9KSlYPvXPEoolUQHnqZ) \cr
#Example dataset from Ontario College Carball Winter 2025. \cr
Result <- AccountCombiner(OCCBStats_noname, main = "5ecc6af45f7c487d8548535e20977db6", alt1 = "76561198799216404")\cr
#Combining accounts from a player who played one series on a seperate account.
}
