\name{RankedStats}
\alias{RankedStats}
\title{
Ranked Stats
}
\description{
Takes a RunReplays dataframe that was created with ranked matches and cleans it up + summarizes it.
}
\usage{
RankedStats(results, TopPlayerOnly = TRUE, RemoveNextUp = FALSE, rank = "Champion")
}
\arguments{
  \item{results}{
The dataframe generated by RunReplays (required).
}
  \item{TopPlayerOnly}{
Only return the player found most frequently in replays. Default - TRUE
}
  \item{RemoveNextUp}{
Remove the second-most common player, useful for if the player you're analyzing mainly plays with one other person. Default - FALSE \cr
ONLY USE THIS IF TopPlayerOnly = TRUE
}
  \item{rank}{
Average stats for each rank, according to Ballchasing. Default - "Champion"\cr
ONLY USE THIS IF TopPlayerOnly = TRUE \cr
Rank options: \cr
Bronze \cr
Silver \cr
Gold \cr
Platinum \cr
Diamond \cr
Champion \cr
GC \cr
Pros \cr

}
}
\value{
Returns a dataframe of overall stats from the input replays.
}
\references{
Ballchasing API Documentation- https://ballchasing.com/doc/api \cr
Thanks to CantFly for creating Ballchasing.
}
\author{
Brett Jacot
}
